requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
opencv-python==4.10.0.84
numpy==1.26.4
pillow==10.1.0
torch==2.5.0
torchvision==0.20.0
facenet-pytorch==2.5.3